+++ src/providers/mod.rs
@@ -169,6 +169,7 @@ impl core::str::FromStr for Provider {
 }
 
 /// The error type for [`Provider`]'s [`core::str::FromStr`] implementation.
+#[must_use]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct ProviderParseError;